scripts/bot_ofertas/requirements.txt
python-telegram-bot
google-search-results
beautifulsoup4
httpx
pydantic
requests
fake-useragent
